move SPARQL queries, see #128
